+++ frontend/public/changelog.md
@@ -1,5 +1,9 @@
 # Riwayat Perubahan
 
+## 2026-03-08 Minggu
+
+- Membersihkan data glosarium yang berawalan angka
+
 ## 2026-03-07 Sabtu
 
 - Mengujicobakan gim pilih ganda di beranda